Icons styles updates
